Allowing HdpH to init before runParIO.
This change allows us to initialise HdpH on our own time and make it
easier to time the "true" length of computations. Note that the minute
we initilise HdpH work-stealing starts - In general you don't want a
long delay between init and runParIO for this reason.

In the future such a change makes it possible to have a re-entrant HdpH calls.